chore: Add docs, reorganize files
